refactor: remove /dev/ prefix and create clean navigation
- Move all sections from (dev) route group to app root for clean URLs
  - /configuration (was /dev/configuration)
  - /generators (was /dev/generators)
  - /performance (was /dev/performance)
  - /rate-limiting (was /dev/rate-limiting)
  - All other dev sections now at root level

- Replace Next.js default homepage with section navigation
  - Clean button layout for all features
  - Status indicators (Active/Coming Soon)
  - Professional styling with Tailwind

- Update all /dev/ path references throughout codebase
  - Auth service URLs and redirects
  - API key management revalidation paths
  - Analytics performance detection
  - Middleware route permissions

- Fix test imports for new auth component locations
  - Updated 3 failing test files to use new (auth)/auth/ paths
  - All tests now passing
