+++ app/(auth)/reset-password/page.tsx
@@ -79,7 +79,7 @@ export default function ResetPasswordPage() {
             </p>
             <div className="mt-6">
               <a
-                href="/dev/auth"
+                href="/auth"
                 className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
               >
                 Back to Auth
@@ -111,7 +111,7 @@ export default function ResetPasswordPage() {
             </p>
             <div className="mt-6">
               <a
-                href="/dev/auth"
+                href="/auth"
                 className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
               >
                 Sign In Now
