+++ app/debug-error/page.tsx
