+++ app/minimal-config/page.tsx
