+++ app/page.tsx
@@ -1,103 +1,152 @@
-import Image from "next/image";
+import Link from 'next/link'
 
-export default function Home() {
+const sections = [
+  {
+    number: 1,
+    name: 'Configuration & API Management',
+    href: '/configuration',
+    status: 'active',
+    description: 'Manage API keys with mock fallback support',
+    features: ['API key CRUD', 'Service validation', 'Mock mode', 'Database storage'],
+  },
+  {
+    number: 2,
+    name: 'Authentication & User Management',
+    href: '/auth',
+    status: 'active',
+    description: 'Auth.js v5 with user CRUD operations',
+    features: ['Email/password auth', 'OAuth providers', 'User profiles', 'Session management'],
+  },
+  {
+    number: 3,
+    name: 'Payments & Billing',
+    href: '/payments',
+    status: 'coming-soon',
+    description: 'Stripe integration with subscriptions',
+    features: ['Checkout flow', 'Subscription management', 'Webhooks', 'Invoice history'],
+  },
+  {
+    number: 4,
+    name: 'Email System',
+    href: '/email',
+    status: 'coming-soon',
+    description: 'Resend + React Email templates',
+    features: ['Template gallery', 'Preview UI', 'Send testing', 'Email logs'],
+  },
+  {
+    number: 5,
+    name: 'AI Styling System',
+    href: '/ai-styling',
+    status: 'coming-soon',
+    description: 'Screenshot to design generation',
+    features: ['Image upload', 'Style extraction', 'Theme generation', 'Code export'],
+  },
+  {
+    number: 6,
+    name: 'AI Site Assistant',
+    href: '/ai-assistant',
+    status: 'coming-soon',
+    description: 'Contextual help system',
+    features: ['Chat interface', 'Context awareness', 'Help articles', 'User feedback'],
+  },
+  {
+    number: 7,
+    name: 'Code Generators',
+    href: '/generators',
+    status: 'active',
+    description: 'Developer productivity tools',
+    features: ['Component templates', 'API endpoints', 'Database models', 'Test scaffolding'],
+  },
+  {
+    number: 8,
+    name: 'Rate Limiting Dashboard',
+    href: '/rate-limiting',
+    status: 'active',
+    description: 'Monitor and manage API rate limiting',
+    features: ['Live statistics', 'Algorithm monitoring', 'Active limits', 'Alert system'],
+  },
+  {
+    number: 9,
+    name: 'Performance Monitor',
+    href: '/performance',
+    status: 'active',
+    description: 'Monitor and optimize application performance',
+    features: ['Core Web Vitals', 'Runtime metrics', 'Resource timings', 'Performance insights'],
+  },
+  {
+    number: 10,
+    name: 'Deployment & CI/CD',
+    href: '/deployment',
+    status: 'coming-soon',
+    description: 'GitHub Actions and Vercel setup',
+    features: ['Auto-deploy', 'Environment config', 'Health checks', 'Monitoring'],
+  },
+]
+
+export default function HomePage() {
   return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <main className="flex flex-col gap-[32px] row-start-2 items-center sm:items-start">
-        <Image
-          className="dark:invert"
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol className="list-inside list-decimal text-sm/6 text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
-          <li className="mb-2 tracking-[-.01em]">
-            Get started by editing{" "}
-            <code className="bg-black/[.05] dark:bg-white/[.06] px-1 py-0.5 rounded font-[family-name:var(--font-geist-mono)] font-semibold">
-              app/page.tsx
-            </code>
-            .
-          </li>
-          <li className="tracking-[-.01em]">
-            Save and see your changes instantly.
-          </li>
-        </ol>
-
-        <div className="flex gap-4 items-center flex-col sm:flex-row">
-          <a
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:w-auto"
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className="dark:invert"
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 w-full sm:w-auto md:w-[158px]"
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Read our docs
-          </a>
+    <div className="min-h-screen bg-gray-50">
+      <div className="max-w-6xl mx-auto px-4 py-12">
+        <div className="text-center mb-12">
+          <h1 className="text-4xl font-bold text-gray-900 mb-4">
+            Modern SaaS Starter Template
+          </h1>
+          <p className="text-xl text-gray-600">
+            Test and explore each feature of the modern SaaS template
+          </p>
         </div>
-      </main>
-      <footer className="row-start-3 flex gap-[24px] flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
+
+        <div className="grid gap-6 md:grid-cols-2">
+          {sections.map((section) => (
+            <Link
+              key={section.number}
+              href={section.status === 'active' ? section.href : '#'}
+              className={`block p-6 bg-white rounded-lg border-2 transition-all ${
+                section.status === 'active'
+                  ? 'border-blue-500 hover:shadow-lg cursor-pointer hover:border-blue-600'
+                  : 'border-gray-200 opacity-60 cursor-not-allowed'
+              }`}
+            >
+              <div className="flex items-start justify-between mb-4">
+                <div>
+                  <span className="text-sm font-medium text-gray-500">
+                    Section {section.number}
+                  </span>
+                  <h2 className="text-xl font-semibold mt-1 text-gray-900">
+                    {section.name}
+                  </h2>
+                </div>
+                <span
+                  className={`px-3 py-1 text-sm font-medium rounded-full ${
+                    section.status === 'active'
+                      ? 'bg-green-100 text-green-700'
+                      : 'bg-gray-100 text-gray-500'
+                  }`}
+                >
+                  {section.status === 'active' ? 'Ready' : 'Coming Soon'}
+                </span>
+              </div>
+              
+              <p className="text-gray-600 mb-4">{section.description}</p>
+              
+              <div className="space-y-1">
+                {section.features.map((feature, i) => (
+                  <div key={i} className="flex items-center text-sm text-gray-500">
+                    <span className="mr-2">•</span>
+                    {feature}
+                  </div>
+                ))}
+              </div>
+            </Link>
+          ))}
+        </div>
+
+        <div className="mt-12 text-center text-gray-500">
+          <p className="text-sm">
+            Click on any "Ready" section to explore its features
+          </p>
+        </div>
+      </div>
     </div>
-  );
+  )
 }
